Voice_Chatbot/requirements.txt

streamlit
openai
gtts
speechrecognition
pydub
python-dotenv
